Add another mock game utility
